requirements.txt
streamlit==1.29.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
pillow==10.1.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
plotly==5.18.0
streamlit-option-menu==0.3.6